modify: 예약 시 시간을 timeId로 지정하도록 API 스펙 수정
