+++ src/main/java/roomescape/domain/reservation/service/request/ReserveRequest.java
@@ -5,7 +5,6 @@ import lombok.Getter;
 import roomescape.domain.reservation.exception.ReservationException;
 
 import java.time.LocalDate;
-import java.time.LocalTime;
 import java.util.Objects;
 
 @Getter
@@ -13,16 +12,16 @@ public class ReserveRequest {
 
     private String name;
     private LocalDate date;
-    private LocalTime time;
+    private Long timeId;
 
     public ReserveRequest() {
     }
 
     @Builder
-    private ReserveRequest(final String name, final LocalDate date, final LocalTime time) {
+    private ReserveRequest(final String name, final LocalDate date, final Long timeId) {
         this.name = name;
         this.date = date;
-        this.time = time;
+        this.timeId = timeId;
     }
 
     public void validateAllFieldsExist() {
@@ -32,7 +31,7 @@ public class ReserveRequest {
         if (Objects.isNull(this.date)) {
             throw ReservationException.nullField("date");
         }
-        if (Objects.isNull(this.time)) {
+        if (Objects.isNull(this.timeId)) {
             throw ReservationException.nullField("time");
         }
     }
