+++ src/test/java/roomescape/domain/reservation/api/ReservationApiTest.java
@@ -1,0 +1,64 @@
+package roomescape.domain.reservation.api;
+
+import io.restassured.RestAssured;
+import io.restassured.http.ContentType;
+import io.restassured.response.ValidatableResponse;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import roomescape.domain.reservationtime.model.ReservationTime;
+import roomescape.domain.reservationtime.repository.ReservationTimeRepository;
+import roomescape.support.RestAssuredTestSupport;
+
+import java.time.LocalDateTime;
+import java.time.LocalTime;
+import java.util.HashMap;
+import java.util.Map;
+
+import static org.hamcrest.Matchers.is;
+
+class ReservationApiTest extends RestAssuredTestSupport {
+
+    @Autowired
+    ReservationTimeRepository timeRepository;
+
+    @Test
+    void reservation() {
+        final ReservationTime time = ReservationTime.builder()
+                .startAt(LocalTime.of(12, 0))
+                .createdAt(LocalDateTime.of(2024, 6, 23, 7, 0))
+                .build();
+
+        final ReservationTime savedTime = timeRepository.save(time);
+
+        Map<String, String> params = new HashMap<>();
+        params.put("name", "브라운");
+        params.put("date", "2023-08-05");
+        params.put("timeId", String.valueOf(savedTime.getIdValue()));
+
+        final ValidatableResponse response = RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .body(params)
+                .when().post("/reservations")
+                .then().log().all()
+                .statusCode(200);
+
+        final Long reservationId = ((Integer) response.extract().path("id")).longValue();
+
+        RestAssured.given().log().all()
+                .when().get("/reservations")
+                .then().log().all()
+                .statusCode(200)
+                .body("size()", is(1));
+
+        RestAssured.given().log().all()
+                .when().pathParam("reservationId", reservationId).delete("/reservations/{reservationId}")
+                .then().log().all()
+                .statusCode(200);
+
+        RestAssured.given().log().all()
+                .when().get("/reservations")
+                .then().log().all()
+                .statusCode(200)
+                .body("size()", is(0));
+    }
+}
